Added a Hamming decoder function. Also made some improvements to the matrix multiplier in the utility file.
As far as I can tell, the decoder works (I did both inspection and random tests), but that's not a great guarantee.

As with the Hamming encoder, if many repeated calls to the function are necessary, it can be made more efficient using a class structure so that the parity check matrix need not be repeatedly generated at each call.

So much for getting to bed on time. At least I got it done though.