Added ability to changed frequency independent of bit rate when using phase modulation.
The solution is not particularly elegant (because it just adds another optional argument, but I think this is better than the alternative I have in mind, because it still forces the use of the modulation scheme chosen, which could be useful if more complex schemes are implemented in the future.
I also changed some of the default arguments to perform some implicit error-checking rather than implement actual error checking myself because I'm lazy.